move readFile to File.h
